backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-nmap==0.7.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
redis==5.0.1
celery==5.3.4
reportlab==4.0.4
jinja2==3.1.2
pdfkit==1.0.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0